+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Fix usage of removed type `_typeshed.SupportsLessThan` (#421)
 - Remove old configuration abstraction (#414)
 
 ## Version 0.6.0 (January 12, 2022)